Unify calendar cell structure with master and move fonts
