+++ src/calendar/tests/functional/Calendar.ts
@@ -4,7 +4,6 @@ const { assert } = intern.getPlugin('chai');
 import { Remote } from 'intern/lib/executors/Node';
 import keys from '@theintern/leadfoot/keys';
 import * as css from '../../../theme/calendar/calendar.m.css';
-import * as cssCell from '../../../theme/calendar/calendarCell.m.css';
 
 const DELAY = 500;
 
@@ -127,7 +126,7 @@ registerSuite('Calendar', {
 			.findByCssSelector(disabledDateSelector)
 				.getAttribute('class')
 				.then((className: string) => {
-					assert.include(className, cssCell.inactiveDate, 'Disabled date has correct css class');
+					assert.include(className, css.inactiveDate, 'Disabled date has correct css class');
 				});
 	},
 
@@ -208,7 +207,7 @@ registerSuite('Calendar', {
 	'Clicking disabled date moves focus'() {
 		let clickedDate = '';
 		return clickDate(this.remote)
-			.findByCssSelector(`.${cssCell.inactiveDate}`)
+			.findByCssSelector(`.${css.inactiveDate}`)
 				.getVisibleText()
 				.then(text => {
 					clickedDate = text;
@@ -223,7 +222,7 @@ registerSuite('Calendar', {
 				})
 				.getAttribute('class')
 				.then((className: string) => {
-					assert.include(className, cssCell.selectedDate, 'Clicked date has selected class');
+					assert.include(className, css.selectedDate, 'Clicked date has selected class');
 				});
 	}
 });
