+++ src/theme/calendar/calendar.m.css
@@ -9,6 +9,27 @@
 	box-sizing: inherit;
 }
 
+.inactiveDate {
+	background-color: var(--disabled-color);
+}
+
+.selectedDate {
+	background-color: var(--selected-background);
+	color: var(--selected-color);
+}
+
+.todayDate {}
+
+.date {
+	cursor: pointer;
+	padding: 10px 20px;
+	text-align: center;
+}
+
+.date:hover {
+	background-color: var(--disabled-color); /* TODO: better hover color variable? */
+}
+
 /* Date grid */
 .dateGrid {}
