+++ src/theme/calendar/calendar.m.css.d.ts
@@ -1,4 +1,8 @@
 export const root: string;
+export const inactiveDate: string;
+export const selectedDate: string;
+export const todayDate: string;
+export const date: string;
 export const dateGrid: string;
 export const weekday: string;
 export const datePicker: string;
