--- src/theme/calendar/calendarCell.m.css.d.ts
@@ -1,4 +1,0 @@
-export const inactiveDate: string;
-export const selectedDate: string;
-export const todayDate: string;
-export const date: string;
