+++ src/theme/common/fonts/dojo2Selection.json
